uapi: convert net/ipv4/ipv4.h
